adjusted some themes, added mattedesktop theme, reset starship prompt

+++ dot_config/scripts/theme-switcher/executable_switch_theme.sh
@@ -6,7 +6,7 @@ update_files() {
   for file in $(find "$theme_dir" -type f -printf '%P\n'); do
     ln -sf "$theme_dir/$file" "$HOME/.config/$file"
 
-    if [[ "$quiet" != 1 ]]; then
+    if [[ $? -eq 0 && "$quiet" != 1 ]]; then
       echo "Created symlink: $theme_dir/$file -> $HOME/.config/$file"
     fi
   done
@@ -27,10 +27,32 @@ execute_global_commands() {
       -e "s|\$wallpaper|$wallpaper|g" \
       -e "s|\$theme_name|$theme_name|g")
 
-    echo "Executing: $expanded_command"
+    if [[ "$quiet" != 1 ]]; then
+      echo "Executing: $expanded_command"
+    fi
 
     eval "$expanded_command"
   done <<<"$GLOBAL_COMMANDS"
+}
+
+# Function to backup existing files
+backup_files() {
+  local theme_dir=$1
+  local backup_dir="$HOME/.cache/theme-switcher/$(date '+%Y-%m-%d_%H:%M:%S')"
+
+  mkdir -p "$backup_dir"
+  echo "Backing up existing files to: $backup_dir"
+
+  for file in $(find "$theme_dir" -type f -printf '%P\n'); do
+    local target_file="$HOME/.config/$file"
+    if [[ -f "$target_file" ]]; then
+      mkdir -p "$(dirname "$backup_dir/$file")"
+      cp "$target_file" "$backup_dir/$file"
+      if [[ "$quiet" != 1 ]]; then
+        echo "Backed up: $target_file -> $backup_dir/$file"
+      fi
+    fi
+  done
 }
 
 # Function to switch themes
@@ -51,13 +73,38 @@ switch_theme() {
   theme_wallpaper="$theme_dir/wallpaper.png"
   theme_config_dir="$theme_dir"
 
+  if [[ "$backup" == 1 ]]; then
+    backup_files "$theme_config_dir"
+    if [[ "$quiet" != 1 ]]; then
+      echo ""
+    fi
+    echo "=> Backup completed"
+  fi
+
+  # List files that will be overwritten:
+  echo "The following files will be modified:"
+  for file in $(find "$theme_dir" -type f -printf '%P\n'); do
+    echo "$HOME/.config/$file"
+  done
+
+  # Ask for user confirmation
+  read -p "Do you want to continue? (y/n) " confirm
+  if [[ "$confirm" != "y" ]]; then
+    echo "Aborted."
+    exit 0
+  fi
+
   update_files "$theme_config_dir"
+  if [[ "$quiet" != 1 ]]; then
+    echo ""
+  fi
   echo "=> Updated config files"
-  echo ""
 
   execute_global_commands "$theme_wallpaper" "$theme_name"
+  if [[ "$quiet" != 1 ]]; then
+    echo ""
+  fi
   echo "=> Executed global commands."
-  echo ""
 }
 
 # Create default config directory if not present
@@ -82,13 +129,14 @@ if [[ $# -eq 0 ]]; then
   echo "Usage: $0 [OPTIONS] <theme_name>"
   echo ""
   echo "Options:"
-  echo "  -l, --list       List all available themes in ~/.config/theme-switcher/themes"
+  echo "  -l, --list       (alone) List all available themes in ~/.config/theme-switcher/themes"
   echo "  -q, --quiet      Suppress output for symlink creation and command execution"
+  echo "  -q, --backup     Backup existing config files"
   echo "  <theme_name>     Name of the theme to switch to"
   echo ""
   echo "Examples:"
   echo "  $0 -l             # List available themes"
-  echo "  $0 -q <theme_name> # Switch to <theme_name> silently (without output)"
+  echo "  $0 -q -b <theme_name> # Backup files, and switch to <theme_name> silently (without output)"
   echo "  $0 <theme_name>    # Switch to <theme_name> and show output"
   exit 1
 fi
@@ -99,9 +147,13 @@ while [[ $# -gt 0 ]]; do
   -l | --list)
     echo "Themes found in ~/.config/theme-switcher/themes:"
     ls -1 ~/.config/theme-switcher/themes/
+    exit 0
     ;;
   -q | --quiet)
     quiet=1
+    ;;
+  -b | --backup)
+    backup=1
     ;;
   *)
     switch_theme "$1"
